Add in SSH agent test and link
We have had feedback that some people are getting errors when running the Ansible bit of the finish script. If the SSH agent isn't started there will be difficulty connecting to the other nodes. This patch adds in a test that users can run and links to the GitHub documentation in agent forwarding.